Module floccus::relative_humidity

Functions to calculate relative humidity in %/100

Functions

general1

Formula for computing relative humidity from mixing ratio and saturation mixing ratio. Can be used interchangeably with general2.

general2

Formula for computing relative humidity from vapour pressure and saturation vapour pressure. Can be used interchangeably with general1.

general3

Formula for computing relative humidity from temperature and dewpoint using tetens1 function for vapour pressure calculation

general4

Formula for computing relative humidity from temperature, dewpoint and pressure using buck3 function for vapour pressure calculation

general5

Formula for computing relative humidity from temperature, dewpoint and pressure using accuracy1 function for mixing ratio calculation