Module floccus::vapour_pressure

Functions to calculate partial vapour pressure of the unsaturated air in Pa.

To compute saturation vapour pressure input dry-bulb temperature in place of dewpoint temperature.

Functions

buck1

Formula for computing vapour pressure from dewpoint temperature and pressure. Should be used for air over water when accuracy is desired.

buck2

Formula for computing vapour pressure from dewpoint temperature and pressure. Should be used for air over ice when accuracy is desired.

buck3

Formula for computing vapour pressure from dewpoint temperature and pressure. Should be used for air over water for general use.

buck3_simplified

Formula for computing vapour pressure from dewpoint temperature. Simplified version of buck3. Very popular in meteorological sources.

buck4

Formula for computing vapour pressure from dewpoint temperature and pressure. Should be used for air over ice for general use.

buck4_simplified

Formula for computing vapour pressure from dewpoint temperature. Simplified version of buck4, analogical to buck3_simplified.

general1

Formula for computing vapour pressure from specific humidity and pressure. This function is theoretical not empirical.

saturation_specific1

Formula for computing ONLY vapour pressure from saturation vapour pressure and relative humidity. For saturation vapour pressure use saturation_specific2

saturation_specific2

Formula for computing ONLY saturation vapour pressure from vapour pressure and relative humidity. For vapour pressure use saturation_specific1

tetens1

Formula for computing vapour pressure over water from dewpoint temperature. Should be used for temperatures above 273K.

wexler1

Formula for computing vapour pressure over water from dewpoint temperature. Should be used when accuracy is required as it is computationally expensive.

wexler2

Formula for computing vapour over ice pressure from dewpoint temperature. Should be used when accuracy is required as it is computationally expensive.