Module floccus::vapour_pressure_deficit

Functions to calculate vapour pressure deficit in Pa.

Vapour-pressure deficit, is the difference (deficit) between the amount of moisture in the air and how much moisture the air can hold when it is saturated (Wikipedia).

Functions

general1

Formula for computing vapour pressure deficit from vapour pressure and saturation vapour pressure

general2

Formula for computing vapour pressure deficit from temperature, dewpoint and pressure using buck3 function for vapour pressure calculation

general3

Formula for computing vapour pressure deficit from temperature, relative humidity and pressure using buck3 function for vapour pressure calculation