Module floccus::virtual_temperature

Functions to calculate virtual temperature of air in K.

In atmospheric thermodynamics, the virtual temperature of a moist air parcel is the temperature at which a theoretical dry air parcel would have a total pressure and density equal to the moist parcel of air (Wikipedia).

Functions

general1

Formula for computing virtual temperature from temperature and mixing ratio.

general2

Formula for computing virtual temperature from air temperature, pressure and vapour pressure.

general3

Formula for computing virtual temperature from air temperature and specific humidity.